:recycle: 사이드탭 컴포넌트 타입, 스타일 분리
