+++ src/components/Filter/SideTab/category.type.ts
@@ -1,0 +1,8 @@
+export type CategoryProps = {
+  categoryLists: { id: number; name: string }[];
+  title: string;
+};
+
+export type CategoryListProps = {
+  isDisplay: boolean;
+};
